pyproject.toml
[tool.poetry]
name = "flask-todos"
version = "0.1.0"
description = "Flask Todos"
authors = ["Zhen Jun Xu <me@arlenx.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.991"
yapf = "^0.32.0"
toml = "^0.10.2"
pre-commit = "^2.20.0"
gunicorn = "^20.1.0"
types-psycopg2 = "^2.9.21.2"
flake8-quotes = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 2
split_before_logical_operator = true

[tool.isort]
profile = "hug"

[tool.mypy]
check_untyped_defs = true
